Update patch set 1
Patch Set 1:

Cristiano: it has nothing to do with any particular theme, but with the new engine. And a lot of people have tried with clean installs and the problem persists. The problem has been confirmed and there are logs. There is definitely a problem somewhere with the new theme engine.

Patch-set: 1