Update patch set 1
Patch Set 1:

You may be right, but this is not the way to handle this. If we would revert any feature that had bugs in the beggining we wouldn't have half the features we have today! Reverting is an easy solution, but that's not why we all here. We are here to develop, not stick in place. And if you have problems, you can just download one of the 5 or 6 milestone releases that came before the new theme engine. If you want nightlies, then you might face problems. Thats why it called nightlies.

Patch-set: 1