Update patch set 1
Patch Set 1:

I just wanted to take a brief moment to let those of you watching this gerrit patch know that I was able to reproduce this problem on my Nexus 5 running the M5 snapshot which uses the OLD theme engine.  So far it appears that this is not a problem introduced with the new theme engine.  The problem may be manifesting itself sooner with the new engine but I do not believe it is a unique issue with the new theme engine.  SystemUI has a huge memory leak in both engines which I have been working on resolving.

At this point I think it is time to end this show and move on to other pressing issues.  For those that are curious, here is the log showing the ANR occurring in SystemUI.
http://logcat.scheffsblend.com/view?id=2519001

Patch-set: 1